Specs: Add File to Experience + Standard Controller Tests
**What**
This approach will test our file uploads for each of the objects we're including them with - eventually we can use this for all the main objects (I've left the tests commented out for now waiting for us to complete the entire suite of attachments available on each object). Once they're all merged, we should create another PR to uncomment the standard controller action tests.